# target/classes/swagger/Mortgage-Api.yaml
openapi: 3.0.3
info:
  title: Mortgage API
  description: API to facilitate operations on mortgages like fetching mortgage rates, calculation eligibility etc
  version: 1.0.0

paths:
  /api/interest-rates:
    get:
      summary: Get current interest rates
      operationId: getInterestRates
      tags:
        - Rates
      responses:
        '200':
          description: List of current interest rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InterestRate'

        '500':
          description: Server error / Operation error
          content:
            application/json:
              schema:
                $ref: "#/definitions/Error"

  /api/mortgage-check:
    post:
      summary: Check mortgage eligibility
      operationId: checkMortgage
      tags:
        - Mortgage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgageRequest'
      responses:
        '200':
          description: Mortgage eligibility result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MortgageResponse'
        '400':
          description: Invalid input / validation errors
          content:
            application/json:
              schema:
                $ref: "#/definitions/Error"
        '500':
          description: Server error / Operation error
          content:
            application/json:
              schema:
                $ref: "#/definitions/Error"

components:
  schemas:
    InterestRate:
      type: object
      properties:
        maturityPeriod:
          type: integer
          example: 10
        interestRate:
          type: number
          example: 7.8
        lastUpdated:
          type: string
          format: date-time

    MortgageRequest:
      type: object
      required:
        - income
        - maturityPeriod
        - loanValue
        - homeValue
      properties:
        income:
          type: number
          minimum: 500
          example: 1000
        maturityPeriod:
          type: integer
          minimum: 1
          maximum: 30
          example: 10
        loanValue:
          type: number
          minimum: 5000
          maximum: 5000000
          example: 45000
        homeValue:
          type: number
          example: 450000

    MortgageResponse:
      type: object
      properties:
        eligible:
          type: boolean
          example: true
        monthlyCosts:
          type: number
          example: 1500

    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string